requirements.txt
psycopg2-binary
django
djangorestframework
Pillow
drf-nested-routers
django-rest-framework-roles
django-cors-headers